Moved tests to integration:
- Tests have more to do with interaction between each other than testing the view model itself
