+++ BalancedFieldLength/Application/test/Application.BalancedFieldLength.Test/Application.BalancedFieldLength.Test.csproj
@@ -20,6 +20,7 @@
     <Compile Include="..\..\..\Core\src\Core.Common.Version\GlobalAssemblyInfo.cs">
       <Link>Properties\GlobalAssemblyInfo.cs</Link>
     </Compile>
+    <Compile Include="MainViewModelIntegrationTest.cs" />
     <Compile Include="Views\TabViews\AircraftDataTabViewModelTest.cs" />
     <Compile Include="Views\TabViews\EngineSettingsTabViewModelTest.cs" />
     <Compile Include="Views\TabViews\GeneralSimulationSettingsTabViewModelTest.cs" />
